Renamed coords module to coordinates
